+++ source/dtk/types/EventForm.d
@@ -9,6 +9,6 @@ struct EventForm
 	Event* event;
 	WidgetI focusedWidget;
 	WidgetI mouseFocusedWidget;
-	ulong mouseFocusedWidget_x;
-	ulong mouseFocusedWidget_y;
+	int mouseFocusedWidget_x;
+	int mouseFocusedWidget_y;
 }
